+++ src/pages/api/fetchTransaction.ts
@@ -7,7 +7,7 @@ export async function fetchTransactions(
     let url = `http://localhost:5001/transactions/${addressId}`;
 
     // Add query parameters for pagination
-    if (direction !== "initial" && transaction_index) {
+    if (direction !== "initial" && transaction_index !== undefined) {
       url += `?direction=${direction}&index=${transaction_index}`;
     } else if (direction !== "initial") {
       url += `?direction=${direction}`;
@@ -15,8 +15,10 @@ export async function fetchTransactions(
 
     const response = await fetch(url);
     if (!response.ok) throw new Error("Failed to fetch transactions");
-    console.log(response);
-    return await response.json();
+
+    const data = await response.json();
+    console.log("Raw response data:", data);
+    return data;
   } catch (error) {
     console.error("Error fetching transactions:", error);
     return [];
